misinformation_adk/orchestrator_agent_reel/requirements.txt
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-cloud-texttospeech==2.14.1
google-genai==0.2.2
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
moviepy==1.0.3
pydub==0.25.1
imageio-ffmpeg==0.4.9
feedparser==6.0.10
requests==2.31.0